+++ src/main/java/com/hanghae/finalProject/rest/attendant/controller/PostController.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.attendant.controller;
+
+public class PostController {
+}
