+++ src/main/java/com/hanghae/finalProject/rest/attendant/dto/PostRequestDto.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.attendant.dto;
+
+public class PostRequestDto {
+}
