+++ src/main/java/com/hanghae/finalProject/rest/attendant/model/Post.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.attendant.model;
+
+public class Post {
+}
