+++ src/main/java/com/hanghae/finalProject/rest/attendant/repository/PostRepository.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.attendant.repository;
+
+public class PostRepository {
+}
