+++ src/main/java/com/hanghae/finalProject/rest/follow/controller/LikeController.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.follow.controller;
+
+public class LikeController {
+}
