+++ src/main/java/com/hanghae/finalProject/rest/follow/dto/LikeRequestDto.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.follow.dto;
+
+public class LikeRequestDto {
+}
