+++ src/main/java/com/hanghae/finalProject/rest/follow/model/Like.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.follow.model;
+
+public class Like {
+}
