+++ src/main/java/com/hanghae/finalProject/rest/follow/repository/LikeRepository.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.follow.repository;
+
+public class LikeRepository {
+}
