+++ src/main/java/com/hanghae/finalProject/rest/follow/service/LikeService.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.follow.service;
+
+public class LikeService {
+}
