--- src/main/java/com/hanghae/finalProject/rest/like/controller/LikeController.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.like.controller;
-
-public class LikeController {
-}
