--- src/main/java/com/hanghae/finalProject/rest/like/dto/LikeRequestDto.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.like.dto;
-
-public class LikeRequestDto {
-}
