--- src/main/java/com/hanghae/finalProject/rest/like/model/Like.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.like.model;
-
-public class Like {
-}
