--- src/main/java/com/hanghae/finalProject/rest/like/repository/LikeRepository.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.like.repository;
-
-public class LikeRepository {
-}
