--- src/main/java/com/hanghae/finalProject/rest/like/service/LikeService.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.like.service;
-
-public class LikeService {
-}
