+++ src/main/java/com/hanghae/finalProject/rest/meeting/controller/PostController.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.meeting.controller;
+
+public class PostController {
+}
