+++ src/main/java/com/hanghae/finalProject/rest/meeting/dto/PostRequestDto.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.meeting.dto;
+
+public class PostRequestDto {
+}
