+++ src/main/java/com/hanghae/finalProject/rest/meeting/model/Post.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.meeting.model;
+
+public class Post {
+}
