+++ src/main/java/com/hanghae/finalProject/rest/meeting/repository/PostRepository.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.meeting.repository;
+
+public class PostRepository {
+}
