--- src/main/java/com/hanghae/finalProject/rest/post/controller/PostController.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.post.controller;
-
-public class PostController {
-}
