--- src/main/java/com/hanghae/finalProject/rest/post/dto/PostRequestDto.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.post.dto;
-
-public class PostRequestDto {
-}
