--- src/main/java/com/hanghae/finalProject/rest/post/model/Post.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.post.model;
-
-public class Post {
-}
