--- src/main/java/com/hanghae/finalProject/rest/post/repository/PostRepository.java
@@ -1,4 +1,0 @@
-package com.hanghae.finalProject.rest.post.repository;
-
-public class PostRepository {
-}
