--- src/main/java/com/hanghae/finalProject/rest/post/service/PostService.java
@@ -1,16 +1,0 @@
-package com.hanghae.finalProject.rest.post.service;
-
-public class PostService {
-
-//     @Transactional
-//     public PostResponseDto.createResponse createPost(PostRequestDto postRequestDto) {
-//          User user = SecurityUtil.getCurrentUser();
-//          MultipartFile file = postRequestDto.getFile();
-//          String imgUrl = null;
-//          if(file != null && file.getContentType() != null) {
-//               imgUrl = s3Uploader.upload(file, "postImage");
-//          }
-//          Post post = postRepository.saveAndFlush(new Post(postRequestDto, user.getUsername(), imgUrl));
-//          return new PostResponseDto.createResponse(post, user.getNickname());
-//     }
-}
