+++ src/main/java/com/hanghae/finalProject/rest/report/controller/PostController.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.report.controller;
+
+public class PostController {
+}
