+++ src/main/java/com/hanghae/finalProject/rest/report/dto/PostRequestDto.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.report.dto;
+
+public class PostRequestDto {
+}
