+++ src/main/java/com/hanghae/finalProject/rest/report/model/Post.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.report.model;
+
+public class Post {
+}
