+++ src/main/java/com/hanghae/finalProject/rest/report/repository/PostRepository.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.report.repository;
+
+public class PostRepository {
+}
