+++ src/main/java/com/hanghae/finalProject/rest/review/controller/LikeController.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.review.controller;
+
+public class LikeController {
+}
