+++ src/main/java/com/hanghae/finalProject/rest/review/dto/LikeRequestDto.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.review.dto;
+
+public class LikeRequestDto {
+}
