+++ src/main/java/com/hanghae/finalProject/rest/review/model/Like.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.review.model;
+
+public class Like {
+}
