+++ src/main/java/com/hanghae/finalProject/rest/review/repository/LikeRepository.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.review.repository;
+
+public class LikeRepository {
+}
