+++ src/main/java/com/hanghae/finalProject/rest/review/service/LikeService.java
@@ -1,0 +1,4 @@
+package com.hanghae.finalProject.rest.review.service;
+
+public class LikeService {
+}
